Compute Wald or percentile bootstrapped confidence interval

Internally calls compute_wald_ci() and/or compute_percentile_ci() depending on the value of ci_type

Usage

compute_boot_ci(
  x,
  boot_x,
  ci_type,
  alpha = 0.05,
  transform = NULL,
  z_star = NULL
)

compute_wald_ci(x, boot_x, alpha = 0.05, transform, z_star = NULL)

compute_percentile_ci(boot_x, alpha = 0.05)

Arguments

x

A numeric vector of point estimates

boot_x

A matrix of bootstrapped estimates where the rows are the bootstrap iterations and the columns are the time points of interest.

ci_type

Character string indicating which type of confidence interval to return ("wald", "percentile", "both")

alpha

Significance level used to compute confidence intervals. Confidence intervals have nominal level 1 - alpha.

transform

If ci_type = "wald", a character string indicating the scale on which to compute Wald confidence intervals that are then transformed back to the original scale. Options are logit for the transformation log(x/(1-x)) or log_ve for the transformation log(1-x).

z_star

If ci_type = "wald", a specific critical value used to to compute Wald confidence intervals (assumed to be positive). If used, alpha argument is ignored.

Value

A matrix containing the lower and upper confidence intervals and relevant bootstrap standard errors.