Get the marginal cumulative incidence in the treated and untreated groups based on a Cox model or Kaplan Meier curve.

#' @description Internally calls compute_cox_ve() or compute_km_ve() depending on the value of method

Usage

compute_marginal_ve(
  adata,
  adata_outcome_name,
  adata_event_name,
  adata_trt_name,
  times,
  method,
  adjust = NULL,
  censor_time = max(times),
  separate = TRUE,
  return_models = TRUE
)

Arguments

adata

A data frame that represents the analysis data set of a clinical trial.

adata_outcome_name

Character string specifying the time to event variable in adata. The time should be the time to event from vaccination/matched index date

adata_event_name

Character string specifying the event variable in adata

adata_trt_name

Character string specifying the treatment variable in adata

times

A numeric vector containing the times (relative to time of vaccination) at which to return the cumulative incidence estimates.

method

Character string specifying method for survival estimation ("cox" for cox proportional hazards model or "km" for kaplan meier estimation)

adjust

If method = "cox", a formula or character vector containing the names of covariates in matched_data to adjust for.

censor_time

The time at which vaccinated individuals are censored during model fitting. By default, this is set to max(times).

separate

If TRUE, models are fit in the treated and untreated groups separately.

return_models

Logical value: Should fitted survival models be returned?

Value

A list containing the following:

estimates

A matrix of estimates. The columns of the matrix are the cumulative incidence/VE terms and the rows are the requested time points for evaluation.

models

If return_models = TRUE, the models used to compute risk and VE