Estimate bootstrapped confidence intervals for proposed VE estimator

Estimate bootstrapped confidence intervals for proposed VE estimator

Usage

estimate_ci(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  marginalizing_dist,
  times,
  censor_time,
  tau,
  boot_formula_0,
  boot_formula_1,
  matched_data,
  gp_list,
  ci_type,
  limit_type,
  n_boot,
  pt_est = NULL,
  alpha = 0.05,
  return_boot = TRUE,
  n_cores = 1
)

Arguments

data

A data frame containing pertinent information for VE estimation.

outcome_name

Character string specifying the name of the outcome variable in data. Outcome variable should be a numeric representing follow-up time relative to study start d0.

event_name

Character string specifying the name of the event indicator variable in data. Event indicator should be numeric-valued: 1=event, 0=censored.

trt_name

Character string specifying name of the vaccination indicator variable in data. Vaccination indicator is an indicator of ever receiving vaccine during the follow-up period and should be numeric-valued: 1=yes, 0 = no.

time_name

Character string specifying name of the vaccination time variable in data. Vaccination time should be NA if never received vaccine during follow-up.

adjust_vars

A character vector containing the names of variables in data to adjust for. #' @param times A numeric vector containing the times (relative to time of vaccination) at which to return the cumulative incidence estimates.

marginalizing_dist

Generally, should be set to "observed", the default value, which estimates the marginalizing distributions based on the observed data. For historical purposes, marginalizing_dist can also be set to "matched", which uses the marginalizing distributions of a matched dataset and requires use of the matched_dist_options argument. To providing a fixed choice of marginalizing weight functions, marginalizing_dist should be a list of 2 dataframes named g_dist and p_dist. g_dist must be a dataframe with the columns <time_name> describing the time of vaccination, all variables in , group_name identifying each unique covariate group, and prob for the corresponding probability. p_dist must be a dataframe with all variables in and variable prob for the corresponding probability.

censor_time

The time at which vaccinated individuals are censored during model fitting. By default, this is set to max(times).

tau

The time excluded after vaccination to allow building up of immunity

boot_formula_0

A formula for estimating hazards in unvaccinated group for bootstrap samples (not a model)

boot_formula_1

A formula for estimating hazards in vaccinated group for bootstrap samples (not a model)

matched_data

A data frame representing the matched cohort

gp_list

List of two dataframes named g_dist and p_dist representing the marginalizing distributions to use.

ci_type

Character string indicating which type of confidence interval to return ("wald", "percentile", "both")

limit_type

Generally, should be set to "limit", the default value. For historical purposes, limit_type can also be set to fixed, which treats the marginalizing distributions as fixed. For a prespecified marginalizing_dist, limit_type is always fixed.

n_boot

Number of bootstrap replicated used to compute confidence intervals

pt_est

A matrix of estimates. The columns of the matrix are the cumulative incidence/VE terms and the rows are the requested timepoints for evaluation.

alpha

Significance level used to compute confidence intervals. Confidence intervals have nominal level 1 - alpha.

return_boot

Logical value: Should bootstrap estimates be returned?

n_cores

Number of cores to use to when running bootstrapping procedure. Passed as mc.cores argument to parallel::mclapply. If n_cores > 1, this parallelizes the bootstrapping procedure on Unix-like systems (not available on Windows).

Value

A list containing the following:

ci_estimates

A list of matrices containing the lower and upper confidence interval bounds and bootstrap standard error for each term. When ci_type = "wald", the bootstrapped standard errors on the transformed scale are also included.

n_success_boot

The number of bootstrap samples used to compute confidence interval

boot_samples

A matrix containing estimates from all bootstrap replications. Rows represent bootstrap iterations, columns the term estimated.