Helper to estimate generic bootstrapped confidence interval

This function computes Wald and percentile bootstrapped confidence intervals for the proposed VE estimator.

Usage

estimate_general_ci(
  one_boot_function_name,
  one_boot_args,
  ci_type,
  n_boot,
  pt_est = NULL,
  alpha = 0.05,
  return_boot = TRUE,
  n_cores = 1
)

Arguments

one_boot_function_name

Character string naming the function that computes one bootstrap iteration of estimates

one_boot_args

List of arguments to pass to the function that computes one bootstrap iteration of estimations

ci_type

Character string indicating which type of confidence interval to return ("wald", "percentile", "both")

n_boot

Number of bootstrap replicated used to compute confidence intervals

pt_est

A matrix of estimates. The columns of the matrix are the cumulative incidence/VE terms and the rows are the requested time points for evaluation.

alpha

Significance level used to compute confidence intervals. Confidence intervals have nominal level 1 - alpha.

return_boot

Logical value: Should bootstrap estimates be returned?

n_cores

Number of cores to use to when running bootstrapping procedure. Passed as mc.cores argument to parallel::mclapply. If n_cores > 1, this parallelizes the bootstrapping procedure on Unix-like systems (not available on Windows).

Value

A list containing the following:

ci_estimates

A list of matrices containing the lower and upper confidence interval bounds and bootstrap standard error for each term. When ci_type = "wald", the bootstrapped standard errors on the transformed scale are also included.

n_success_boot

A numeric vector of the number of successful bootstrap samples for each time point.(Success bootstrap samples are those that result in non-missing valid point estimates.)

boot_samples

If return_boot = TRUE, a list of matrices for each term that contain the bootstrap estimates where the rows are the bootstrap iterations and the columns are the time points.