Main function to estimate vaccine effectiveness using a G-computation style estimator

The primary function to estimate vaccine effectiveness.

Usage

obsve(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  marginalizing_dist,
  matched_dist_options = NULL,
  times,
  censor_time = max(times),
  tau = 14,
  ci_type = "wald",
  limit_type = "fixed",
  n_boot = 0,
  alpha = 0.05,
  formula_0 = NULL,
  formula_1 = NULL,
  return_models = TRUE,
  return_boot = TRUE,
  n_cores = 1
)

Arguments

data

A data frame containing pertinent information for VE estimation.

outcome_name

Character string specifying the name of the outcome variable in data. Outcome variable should be a numeric representing follow-up time relative to study start d0.

event_name

Character string specifying the name of the event indicator variable in data. Event indicator should be numeric-valued: 1=event, 0=censored.

trt_name

Character string specifying name of the vaccination indicator variable in data. Vaccination indicator is an indicator of ever receiving vaccine during the follow-up period and should be numeric-valued: 1=yes, 0 = no.

time_name

Character string specifying name of the vaccination time variable in data. Vaccination time should be NA if never received vaccine during follow-up.

adjust_vars

A character vector containing the names of variables in data to adjust for.

marginalizing_dist

Character string describing the type of estimated day/covariate distributions to marginalize over. Values are "observed" or "matched". Alternatively, for pre-specified marginalizing distributions can be a list of 2 dataframes named g_dist and p_dist.

matched_dist_options

If marginalizing_dist == "matched", a list of parameters controlling process of creating matched cohort and matched analysis dataset from which marginalizing distributions will be estimated. List can be created by matched_dist().

times

A numeric vector containing the times (relative to time of vaccination) at which to return the cumulative incidence estimates.

censor_time

The time at which vaccinated individuals are censored during model fitting. By default, this is set to max(times).

tau

The time excluded after vaccination to allow building up of immunity

ci_type

Character string indicating which type of confidence interval to return ("wald", "percentile", "both")

limit_type

Character string indicating whether the marginalizing distributions of interest are the estimated distributions in the "limit" or "fixed" sense. For a prespecified marginalizing_dist, limit_type is always fixed.

n_boot

Number of bootstrap replicated used to compute confidence intervals

alpha

Significance level used to compute confidence intervals. Confidence intervals have nominal level 1 - alpha.

formula_0

A formula or pre-fit model for estimating hazards in unvaccinated group (recommended for use only by power users)

formula_1

A formula or pre-fit model for estimating hazards in vaccinated group. (recommended for use only by power users)

return_models

Logical value: Should fitted survival models be returned?

return_boot

Logical value: Should bootstrap estimates be returned?

n_cores

Number of cores to use to when running bootstrapping procedure. Passed as mc.cores argument to parallel::mclapply. If n_cores > 1, this parallelizes the bootstrapping procedure on Unix-like systems (not available on Windows).

Value

A list. In addition to information related to the call/arguments to obsve, contains the following:

estimates

A list of matrices for cumulative incidences and VE. Each matrix contains the point estimate and confidence intervals for the specified term.

gp_list

A list containing the distributions g(d|x) and p(x) used for marginalization

model_0

The fit object used to predict risk for unvaccinated group

model_1

The fit object used to predict risk for vaccinated group

n_success_boot

A numeric vector of the number of successful bootstrap samples for each time point.(Success bootstrap samples are those that result in non-missing valid point estimates.)

boot_samples

If return_boot = TRUE, a list of matrices for each term that contain the bootstrap estimates where the rows are the bootstrap iterations and the columns are the time points.