Compute one bootstrap replicate of matching-based VE point estimate

Compute one bootstrap replicate of matching-based VE point estimate

Usage

one_boot_matching(
  matched_data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  method,
  adjust,
  times,
  censor_time,
  tau,
  pair_censoring = TRUE,
  separate = FALSE,
  limit_type = "fixed",
  data = NULL,
  id_name = "ID",
  matching_vars = NULL,
  replace = FALSE,
  return_boot = TRUE
)

Arguments

matched_data

A data frame for the matched cohort

outcome_name

Character string specifying the name of the outcome variable in data. Outcome variable should be a numeric representing follow-up time relative to study start d0.

event_name

Character string specifying the name of the event indicator variable in data. Event indicator should be numeric-valued: 1=event, 0=censored.

trt_name

Character string specifying name of the vaccination indicator variable in data. Vaccination indicator is an indicator of ever receiving vaccine during the follow-up period and should be numeric-valued: 1=yes, 0 = no.

time_name

Character string specifying name of the vaccination time variable in data. Vaccination time should be NA if never received vaccine during follow-up.

method

Character string specifying method for survival estimation ("cox" for cox proportional hazards model or "km" for kaplan meier estimation)

adjust

If method = "cox", a formula or character vector containing the names of covariates in matched_data to adjust for.

times

A numeric vector containing the times (relative to time of vaccination) at which to return the cumulative incidence estimates.

censor_time

The time at which vaccinated individuals are censored during model fitting. By default, this is set to max(times).

tau

The time excluded after vaccination to allow building up of immunity

pair_censoring

Logical to indicate if matched pairs should be censored when unvaccinated pair becomes vaccinated.

separate

If TRUE, models are fit in the treated and untreated groups separately.

limit_type

If limit_type = "limit", the matching procedure is performed in each bootstrap iteration. If limit_type = "fixed"`, bootstrap samples are formed from the original matched data set.

data

If limit_type = "limit", the original data used to constructed the matched cohort

id_name

If limit_type = "limit", character string representing the individual identifier variable in data

matching_vars

If limit_type = "limit", a character vector containing the names of variables in data to match on.

replace

If limit_type = "limit", logical: Should matching be done with replacement?

return_boot

Logical value: Should bootstrap estimates be returned?

Value

A matrix of bootstrapped estimates where the the columns of the matrix are the cumulative incidence/VE terms and the rows are the requested time points for evaluation.