Compute one bootstrap replicate of VE point estimate

This function creates a bootstrapped sample and computes the corresponding point estimate

Usage

one_boot_ve(
  data,
  outcome_name,
  event_name,
  trt_name,
  time_name,
  adjust_vars,
  marginalizing_dist,
  times,
  censor_time,
  tau,
  boot_formula_0,
  boot_formula_1,
  matched_data,
  gp_list,
  limit_type
)

Arguments

data

A data frame containing pertinent information for VE estimation.

outcome_name

Character string specifying the name of the outcome variable in data. Outcome variable should be a numeric representing follow-up time relative to study start d0.

event_name

Character string specifying the name of the event indicator variable in data. Event indicator should be numeric-valued: 1=event, 0=censored.

trt_name

Character string specifying name of the vaccination indicator variable in data. Vaccination indicator is an indicator of ever receiving vaccine during the follow-up period and should be numeric-valued: 1=yes, 0 = no.

time_name

Character string specifying name of the vaccination time variable in data. Vaccination time should be NA if never received vaccine during follow-up.

adjust_vars

A character vector containing the names of variables in data to adjust for.

marginalizing_dist

Character string describing the type of estimated day/covariate distributions to marginalize over. Values are "observed" or "matched". Alternatively, for pre-specified marginalizing distributions can be a list of 2 dataframes named g_dist and p_dist.

times

A numeric vector containing the times (relative to time of vaccination) at which to return the cumulative incidence estimates.

censor_time

The time at which vaccinated individuals are censored during model fitting. By default, this is set to max(times).

tau

The time excluded after vaccination to allow building up of immunity

boot_formula_0

A formula for estimating hazards in unvaccinated group for bootstrap samples (not a model)

boot_formula_1

A formula for estimating hazards in vaccinated group for bootstrap samples (not a model)

matched_data

A data frame representing the matched cohort

gp_list

List of two dataframes named g_dist and p_dist representing the marginalizing distributions to use.

limit_type

Character string indicating whether the marginalizing distributions of interest are the estimated distributions in the "limit" or "fixed" sense. For a prespecified marginalizing_dist, limit_type is always fixed.

Value

A matrix of bootstrapped estimates where the the columns of the matrix are the cumulative incidence/VE terms and the rows are the requested time points for evaluation.